applied atomic design and changed meta info
