+++ products/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "products",
-  "version": "0.1.0",
+  "name": "productslandingpage",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
     "axios": "^0.19.0",
