+++ products/src/components/organisms/Products.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import axios from "axios";
-import ProductCard from "./ProductCard";
+import ProductCard from "../molecule/ProductCard";
 
 const Products = () => {
   const [products, setProducts] = useState([]);
